no more repo mixing

+++ prepare_development_parruc.sh
@@ -16,8 +16,11 @@ echo "changing git credentials for your user: $GIT_USER"
 cd "$HOST_DIR/django"
 git config --global user.email "parruc@gmail.com"
 git config --global user.name "Matteo Parrucci"
-git remote rm origin
-git remote add origin https://$GIT_USER@bitbucket.org/flyingfrog/4hm.git
+if `pwd` == "$HOST_DIR/django"; then
+    git remote add origin https://$GIT_USER@bitbucket.org/flyingfrog/4hm.git
+    git remote rm origin
+    git remote add origin https://$GIT_USER@bitbucket.org/flyingfrog/4hm.git
+fi
 
 
 #cd $CWD
